Fix page name not shown when printing
The page name below a source link node is now printed even if the link node is the lowest node in the graph. Before this commit, the page name was cut-off because it was outside the cell bounds. Now, the page name is within the cell bounds (only for the height!)

* Fix: Remove EditorUi from AttackGraphLinkShape